moved raDecFromPupilCoords and pupilCoordsFromRaDec to their own file (FocalPlaneUtils.py)
